+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changes:
 
 Fixes:
 
+- Issue loading significant wave height forecasts due to error in string replacement.
+
 ## 0.5.1 - 9/30/2020
 
 Changes: